fix: reconnect chat socket when internet is available
